All HTML, CSS, and PHP files for the Website
HTML and CSS styling code displays the web page content while all PHP code accesses the database to obtain appropriate data.

All files and the database were created from scratch for the purpose of this project. Intent was to learn how to create a website and database from scratch and have them interact with each other.